+++ app/forms/invite_form.rb
@@ -1,0 +1,6 @@
+class InviteForm < Reform::Form
+  property :email
+  validates :email, :presence => true
+  validates :email, :email => true
+  validates_uniqueness_of :email
+end