+++ config/breadcrumbs.rb
@@ -28,6 +28,10 @@ crumb :category do |name|
   parent :public_node, Node.root_node
 end
 
+crumb :invite do |name|
+  parent :public_node, Node.root_node
+end
+
 crumb :editorial_root do
   link 'Editorial Home', editorial_root_path
 end