+++ config/routes.rb
@@ -96,6 +96,11 @@ Rails.application.routes.draw do
 
   resources :departments, only: :index
   resources :ministers, only: :index
+  resources :invites, only: [:show, :create, :new, :edit, :update] do
+    collection do
+      get :required
+    end
+  end
 
   get 'categories/:slug' => 'categories#show', as: :category